Native search screen and Tab-builder navigation; drop tvOS 17
Search (Classic style):
- Replace the hand-built search bar, hidden UITextField and modal keyboard
  with `.searchable`. tvOS now provides the field, keyboard, dictation and
  live results as you type.
- Remove the type filter chips, the recent-searches row and the Discover
  filter bar. Searching needs none of them; browsing by genre/sort is a
  Home concern, not a search one.
- New `SearchDiscoverGrid`: the empty-state grid (popular titles) built from
  the shared `DiscoverViewModel` and `PosterGridCard`, so both halves of the
  screen use the same cell and captions. `DiscoverSection` is now used only
  by the Netflix-style screen, which is unchanged.
- Grids use flexible adaptive columns centred in their container, so the
  layout fits beside either keyboard style with no trailing gutter.

Navigation:
- Move `TVTab` and `TVMainTabView` out of NuvioTVApp.swift into
  UI/Navigation/ (pure relocation; `private` dropped on the view).
- Declare tabs with the tvOS 18 `Tab` builder and give Search
  `role: .search`, so the sidebar integrates the system search field
  instead of floating the tab pill over it.
- Raise the deployment target from 17.5 to 18.0; the `#available(tvOS 18)`
  fallback in the tab view is gone. tvOS 18 excludes no hardware.
