+++ tvosApp/NuvioTV/Sources/NuvioTVApp.swift
@@ -69,42 +69,6 @@ private enum DetailsReturnDestination {
     case collectionFolder(TVCollectionFolderItem, collectionTitle: String)
 }
 
-enum TVTab: String, CaseIterable, Identifiable {
-    case profile = "Profile"
-    case home = "Home"
-    case search = "Search"
-    case library = "Library"
-    case settings = "Settings"
-
-    var id: String { rawValue }
-
-    /// Localized tab label (rawValue remains stable for selection identity).
-    var title: String {
-        switch self {
-        case .profile:
-            return L10n.string("settings_profiles", fallback: "Profile")
-        case .home:
-            return L10n.string("nav_home", fallback: "Home")
-        case .search:
-            return L10n.string("nav_search", fallback: "Search")
-        case .library:
-            return L10n.string("nav_library", fallback: "Library")
-        case .settings:
-            return L10n.string("nav_settings", fallback: "Settings")
-        }
-    }
-
-    var symbol: String {
-        switch self {
-        case .profile: return "person.crop.circle"
-        case .home: return "house"
-        case .search: return "magnifyingglass"
-        case .library: return "rectangle.stack"
-        case .settings: return "gearshape"
-        }
-    }
-}
-
 /// Main content view - entry point for the app with screen routing
 struct ContentView: View {
     @Environment(\.scenePhase) private var scenePhase
@@ -2314,227 +2278,6 @@ final class ProfileTabAvatarRenderer: ObservableObject {
         // Bail if the active profile changed while the face was downloading.
         guard renderedAvatarId == avatarId else { return }
         image = composed
-    }
-}
-
-private struct TVMainTabView: View {
-    @Binding var selectedTab: TVTab
-    let activeProfile: Profile?
-    @ObservedObject var searchViewModel: SearchViewModel
-    @ObservedObject var netflixSearchViewModel: NetflixSearchViewModel
-    @ObservedObject var libraryViewModel: LibraryViewModel
-    @ObservedObject var homeStore: TVHomeStore
-    let homeCatalogRevision: UInt
-    let homeCollectionsRevision: UInt
-    let isFullScreenOverlayPresented: Bool
-    let detailsDidDisappearGeneration: UInt
-    let accountEmail: String?
-    let isAuthenticated: Bool
-    let sessionNeedsReauthentication: Bool
-    let isProfileSwitching: Bool
-    let authManager: AuthManager
-    let syncManager: NuvioSyncManager
-    let onSwitchProfile: () -> Void
-    let onChangeProfileAvatar: (String, String) -> Void
-    let onChangeProfileName: (String, String) -> Void
-    let onChangeProfilePin: (String, String?, String?) async -> Bool
-    let onVerifyProfilePin: (String, String) async -> Bool
-    let onSignIn: () -> Void
-    let onSignOut: () -> Void
-    let onNavigateToDetails: (String, String) -> Void
-    let onRequestAccountRefresh: () -> Void
-    let onOpenCollectionFolder: (TVCollectionFolderItem, String) -> Void
-    let onResumePlayback: (ContinueWatchingItem) -> Void
-    var onPlayContinueWatchingManually: ((ContinueWatchingItem) -> Void)? = nil
-    var onStartContinueWatchingFromBeginning: ((ContinueWatchingItem) -> Void)? = nil
-    var onRemoveFromContinueWatching: ((ContinueWatchingItem) -> Void)? = nil
-    let onLongPressCard: (NuvioMeta) -> Void
-    /// Long press on a Continue Watching card, which gets its own resume-centric
-    /// menu instead of the generic title actions.
-    let onLongPressContinueWatching: (ContinueWatchingItem) -> Void
-    let onOpenCloudLibrary: () -> Void
-    let onPlayCloudFile: (URL, NuvioMeta) -> Void
-    @AppStorage(SettingsKey.amoled) private var amoled = false
-    @AppStorage(SettingsKey.bodyColor) private var bodyColor = SettingsBackground.charcoal.rawValue
-    @AppStorage(SettingsKey.discoverLocation) private var discoverLocation = "Search"
-    @AppStorage(SettingsKey.searchStyle) private var searchStyle = "Netflix"
-    @AppStorage(SettingsKey.profileName) private var settingsProfileName = "Nuvio User"
-    @StateObject private var profileTabAvatar = ProfileTabAvatarRenderer()
-    @State private var showingReauthSheet = false
-
-    private var displayedProfile: Profile? {
-        if isAuthenticated { return activeProfile }
-        return activeProfile?.id == "guest" ? activeProfile : nil
-    }
-
-    /// Name shown on the profile tab, mirroring the sidebar header's
-    /// display-name logic.
-    private var profileTabTitle: String {
-        guard let displayedProfile else { return "Nuvio Guest" }
-        return ProfileDisplayName.resolve(profile: displayedProfile, settingsName: settingsProfileName)
-    }
-
-    var body: some View {
-        if #available(tvOS 18.0, *) {
-            tabs
-                .tabViewStyle(.sidebarAdaptable)
-        } else {
-            tabs
-        }
-    }
-
-    /// Search screen chosen in Settings → Layout & Discovery → Search Style.
-    @ViewBuilder
-    private var searchTab: some View {
-        if searchStyle == "Classic" {
-            SearchView(
-                viewModel: searchViewModel,
-                showDiscover: discoverLocation == "Search",
-                onContentClick: onNavigateToDetails,
-                onLongPress: onLongPressCard
-            )
-        } else {
-            NetflixSearchView(
-                viewModel: netflixSearchViewModel,
-                showDiscover: discoverLocation == "Search",
-                onContentClick: onNavigateToDetails,
-                onLongPress: onLongPressCard
-            )
-        }
-    }
-
-    private var tabs: some View {
-        TabView(selection: $selectedTab) {
-            // Keep profile switching as a regular tab on every supported tvOS
-            // version. This compiles with the tvOS 26.5 SDK and remains visible
-            // when the app is sideloaded onto tvOS 27.
-            // The tab label carries the profile name + avatar icon so the menu
-            // shows who's signed in instead of a generic "Profile" entry. Its
-            // content stays empty: selecting it goes straight to profile
-            // switching, while editing now lives in Settings.
-            Color.clear
-                .tabItem {
-                    Label {
-                        Text(profileTabTitle)
-                    } icon: {
-                        if sessionNeedsReauthentication {
-                            Image(systemName: "person.crop.circle.badge.exclamationmark")
-                        } else if let avatar = profileTabAvatar.image {
-                            Image(uiImage: avatar).renderingMode(.original)
-                        } else {
-                            Image(systemName: ProfileAvatarCatalog.symbolName(for: displayedProfile?.avatarId))
-                        }
-                    }
-                }
-                .tag(TVTab.profile)
-
-            TVHomeView(
-                store: homeStore,
-                repository: CinemetaCatalogRepository(),
-                isActive: selectedTab == .home,
-                isFullScreenOverlayPresented: isFullScreenOverlayPresented,
-                detailsDidDisappearGeneration: detailsDidDisappearGeneration,
-                isProfileSwitching: isProfileSwitching,
-                contentIdentity: TVHomeContentIdentity(
-                    profileId: activeProfile?.id ?? "none",
-                    catalogRevision: homeCatalogRevision
-                ),
-                collectionsRevision: homeCollectionsRevision,
-                sessionNeedsReauthentication: sessionNeedsReauthentication,
-                onNavigateToDetails: onNavigateToDetails,
-                onOpenCollectionFolder: onOpenCollectionFolder,
-                onResumePlayback: onResumePlayback,
-                onPlayContinueWatchingManually: onPlayContinueWatchingManually,
-                onStartContinueWatchingFromBeginning: onStartContinueWatchingFromBeginning,
-                onRemoveFromContinueWatching: onRemoveFromContinueWatching,
-                onLongPressCard: onLongPressCard,
-                onLongPressContinueWatching: onLongPressContinueWatching,
-                onRequestAccountRefresh: onRequestAccountRefresh,
-                onRequestReauth: { showingReauthSheet = true }
-            )
-                .id(activeProfile?.id ?? "none")
-                .tabItem {
-                    Label(TVTab.home.title, systemImage: TVTab.home.symbol)
-                }
-                .tag(TVTab.home)
-
-            searchTab
-                .tabItem {
-                    Label(TVTab.search.title, systemImage: TVTab.search.symbol)
-                }
-                .tag(TVTab.search)
-
-            LibraryView(
-                viewModel: libraryViewModel,
-                store: ProfileSettings.store(for: activeProfile?.id),
-                onContentClick: onNavigateToDetails,
-                onLongPress: onLongPressCard,
-                onOpenCloudLibrary: onOpenCloudLibrary,
-                onPlayCloudFile: onPlayCloudFile
-            )
-                .id(activeProfile?.id ?? "none")
-                .tabItem {
-                    Label(TVTab.library.title, systemImage: TVTab.library.symbol)
-                }
-                .tag(TVTab.library)
-
-            SettingsView(
-                activeProfile: displayedProfile,
-                accountEmail: accountEmail,
-                isAuthenticated: isAuthenticated,
-                sessionNeedsReauthentication: sessionNeedsReauthentication,
-                onChangeProfileName: onChangeProfileName,
-                onChangeProfileAvatar: onChangeProfileAvatar,
-                onChangeProfilePin: onChangeProfilePin,
-                onVerifyProfilePin: onVerifyProfilePin,
-                onSignIn: {
-                    if sessionNeedsReauthentication {
-                        showingReauthSheet = true
-                    } else {
-                        onSignIn()
-                    }
-                },
-                onSignOut: onSignOut
-            )
-                .tabItem {
-                    Label(
-                        TVTab.settings.title,
-                        systemImage: sessionNeedsReauthentication ? "exclamationmark.circle" : TVTab.settings.symbol
-                    )
-                }
-                .tag(TVTab.settings)
-        }
-        .background(Color.nuvioBackground(amoled: amoled, body: bodyColor).ignoresSafeArea())
-        .sheet(isPresented: $showingReauthSheet) {
-            ReauthSheet(auth: authManager) {
-                syncManager.beginPostLoginSync()
-            }
-        }
-        .onAppear {
-            AvatarCatalogStore.shared.loadIfNeeded()
-            profileTabAvatar.refresh(avatarId: displayedProfile?.avatarId)
-            if sessionNeedsReauthentication {
-                showingReauthSheet = true
-            }
-        }
-        .onChange(of: sessionNeedsReauthentication) { _, needsReauth in
-            if needsReauth {
-                showingReauthSheet = true
-            }
-        }
-        .onChange(of: displayedProfile?.avatarId) { _, newValue in
-            profileTabAvatar.refresh(avatarId: newValue)
-        }
-        .onChange(of: selectedTab) { _, tab in
-            if tab == .profile {
-                onSwitchProfile()
-            }
-        }
-        // Re-attempt once the catalog finishes loading, since the first refresh
-        // can't resolve the avatar image before then.
-        .onReceive(AvatarCatalogStore.shared.$items) { _ in
-            profileTabAvatar.refresh(avatarId: displayedProfile?.avatarId)
-        }
     }
 }
